+++ src/Cocona.Core/ShellCompletion/Candidate/CompletionCandidateResultType.cs
@@ -1,5 +1,8 @@
 namespace Cocona.ShellCompletion.Candidate
 {
+    /// <summary>
+    /// Specifies the type of a completion candidate result.
+    /// </summary>
     public enum CompletionCandidateResultType
     {
         Default,
