+++ src/Cocona.Core/ShellCompletion/ICoconaShellCompletionCodeProvider.cs
@@ -1,47 +1,42 @@
-using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Text;
 using Cocona.Command;
 using Cocona.ShellCompletion.Candidate;
-using Cocona.ShellCompletion.Generators;
 
 namespace Cocona.ShellCompletion
 {
+    /// <summary>
+    /// Provides to generate shell completion code and candidates.
+    /// </summary>
     public interface ICoconaShellCompletionCodeProvider
     {
+        /// <summary>
+        /// Gets the shell names supported by the provider.
+        /// </summary>
         IEnumerable<string> SupportedTargets { get; }
+
+        /// <summary>
+        /// Gets the value that indicates whether the provider can handle.
+        /// </summary>
+        /// <param name="target">The target shell name</param>
+        /// <returns></returns>
         bool CanHandle(string target);
+
+        /// <summary>
+        /// Generates a shell completion support code for the target shell.
+        /// </summary>
+        /// <param name="target">The target shell name</param>
+        /// <param name="writer">The <see cref="TextWriter"/> which to the code is written</param>
+        /// <param name="commandCollection">The <see cref="CommandCollection"/></param>
         void Generate(string target, TextWriter writer, CommandCollection commandCollection);
+
+        /// <summary>
+        /// Generates formatted candidates for the target shell.
+        /// </summary>
+        /// <param name="target">The target shell name</param>
+        /// <param name="writer">The <see cref="TextWriter"/> which to the code is written</param>
+        /// <param name="values">The candidates</param>
         void GenerateOnTheFlyCandidates(string target, TextWriter writer, IReadOnlyList<CompletionCandidateValue> values);
-    }
-
-    public class CoconaShellCompletionCodeProvider : ICoconaShellCompletionCodeProvider
-    {
-        private readonly ICoconaShellCompletionCodeGenerator[] _providers;
-
-        public IEnumerable<string> SupportedTargets { get; }
-
-        public bool CanHandle(string target)
-            => _providers.Any(x => x.Targets.Contains(target));
-
-        public CoconaShellCompletionCodeProvider(IEnumerable<ICoconaShellCompletionCodeGenerator> providers)
-        {
-            _providers = providers?.ToArray() ?? throw new ArgumentNullException(nameof(providers));
-            SupportedTargets = _providers.SelectMany(xs => xs.Targets).ToArray();
-        }
-
-        public void Generate(string target, TextWriter writer, CommandCollection commandCollection)
-        {
-            var provider = _providers.First(x => x.Targets.Contains(target));
-            provider.Generate(writer, commandCollection);
-        }
-
-        public void GenerateOnTheFlyCandidates(string target, TextWriter writer, IReadOnlyList<CompletionCandidateValue> values)
-        {
-            var provider = _providers.First(x => x.Targets.Contains(target));
-            provider.GenerateOnTheFlyCandidates(writer, values);
-        }
     }
 }
